Implemented and tested relative support and confidence adjustment
There is a isSamplingOn flag in StatsComputer and ShapesExtractor files to know if the relative support should be computed or not. 

The relative support is computed based on the percentage of sampled entities.

Confidence of a property is computed based on the entities sampled in the reservoir. 

There was a issue in the printing of csv logs for missing values. This has been fixed now.